requirements.txt
# DSP Eco Tracker - Production Requirements v2.0 (with fake-useragent fix)

# Flask Web Framework
Flask==3.1.0
flask-cors==5.0.1
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
gunicorn==23.0.0
Werkzeug==3.1.3

# Database
PyMySQL==1.1.1
redis==6.0.0

# Machine Learning & Data Science
scikit-learn==1.6.1
xgboost==3.0.0
pandas==2.2.3
numpy==2.2.5
joblib==1.4.2
imbalanced-learn==0.13.0

# Data Visualization (for ML training & analysis)
matplotlib==3.10.0
seaborn==0.13.2

# Web Scraping
selenium==4.31.0
beautifulsoup4==4.13.4
requests==2.32.3
undetected-chromedriver==3.5.5
webdriver-manager==4.0.2
fake-useragent==1.5.1

# Geographic & Utilities  
pgeocode==0.5.0
pycountry>=23.12.11
python-dotenv==1.1.0
python-dateutil==2.9.0

# Security & Authentication
werkzeug>=3.0.0

# Data Processing
marshmallow>=3.20.0

# Optional for enhanced features
altair==5.5.0